♻️(front) rename LayouMainArea.tsx file into index.tsx
We decided to rename every files containing a component into index.tsx
