+++ src/frontend/components/LayoutMainArea/index.tsx
