+++ src/frontend/components/UploadForm/index.tsx
@@ -15,7 +15,7 @@ import { Maybe } from '../../utils/types';
 import { DASHBOARD_ROUTE } from '../Dashboard/route';
 import { ERROR_COMPONENT_ROUTE } from '../ErrorComponent/route';
 import { IframeHeading } from '../Headings';
-import { LayoutMainArea } from '../LayoutMainArea/LayoutMainArea';
+import { LayoutMainArea } from '../LayoutMainArea';
 import { UploadField } from '../UploadField/UploadField';
 
 const messages = defineMessages({
